SGX PVR libraries - Remove the 544_112 libs as we have no use for them on the archos G9.
The gralloc.archos.so was formly a gralloc.omap4430.so from the kindle fire
version of CM-10.2 - The official TI repository was carrying a gralloc which used
libutils for a selection of callstack functions.
The ABI to these functions have changed make the old gralloc incompatible without some
dirty hacks on many fronts.

The "new" gralloc uses libcorkscrew instead so we have a fucking bucket full of win
going on.